Update descriptions.json
Proposed fix to https://github.com/dotnet/corefx/issues/20118. Ideally the text should specify the version of the SDK the package version corresponds to. Can this be done for each version of the package or will it be be an overkill?
The text should then be: 
"Automatically referenced by the .NET Core SDK (version X.x.x). Provides a set of .NET API's that are included in the default .NET Core application model.  "